+++ src/RcsbFv/RcsbFvConfig/RcsbFvConfigInterface.ts
@@ -26,11 +26,11 @@ export interface RcsbFvBoardConfigInterface {
     /**Show tooltip when hovering track annotations*/
     includeTooltip?: boolean;
     /**Function that will be called when a track annotation is clicked*/
-    elementClickCallBack?:(d?:RcsbFvTrackDataElementInterface)=>void;
+    elementClickCallBack?:(d?:RcsbFvTrackDataElementInterface, e?: MouseEvent)=>void;
     /**Function that will be called when hovering a track annotation*/
-    elementEnterCallBack?:(d?:RcsbFvTrackDataElementInterface)=>void;
+    elementEnterCallBack?:(d?:RcsbFvTrackDataElementInterface, e?: MouseEvent)=>void;
     /**Function that will be called when mouse-leaving a track annotation*/
-    elementLeaveCallBack?:(d?:RcsbFvTrackDataElementInterface)=>void;
+    elementLeaveCallBack?:(d?:RcsbFvTrackDataElementInterface, e?: MouseEvent)=>void;
     /**Function that will be called when selection changes*/
     selectionChangeCallBack?:(selection: Array<SelectionInterface>)=>void;
     /**Feature cells border color*/
@@ -62,11 +62,11 @@ interface CommonConfigInterface{
     /**Flag used in sequence display type to force displaying a line when sequence objects are not visible*/
     nonEmptyDisplay?: boolean;
     /**Function that will be called when annotations in this track are clicked*/
-    elementClickCallBack?:(d?:RcsbFvTrackDataElementInterface)=>void;
+    elementClickCallBack?:(d?:RcsbFvTrackDataElementInterface, e?: MouseEvent)=>void;
     /**Function that will be called when clicking annotations in this track*/
-    elementEnterCallBack?:(d?:RcsbFvTrackDataElementInterface)=>void;
+    elementEnterCallBack?:(d?:RcsbFvTrackDataElementInterface, e?: MouseEvent)=>void;
     /**Function that will be called when mouse-leaving a track annotation*/
-    elementLeaveCallBack?:(d?:RcsbFvTrackDataElementInterface)=>void;
+    elementLeaveCallBack?:(d?:RcsbFvTrackDataElementInterface, e?: MouseEvent)=>void;
     /**Show tooltip when hovering annotations in this specific track*/
     includeTooltip?: boolean;
     /**Function that will be called to update track displayed data when the board is moved or zoomed*/